requirements.txt
boto3==1.38.23
python-dotenv==1.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0